Refactored Packets
- Packets now operate on their native byte arrays, no longer on Strings
converted to byte arrays then back.
- Added IDCounter to replace all the counters for next available ID's
and such.
- Moved packet parsing from ClientSocketReceive into ClientGame. This
makes for less synchronization issues, as the receiving thread simply
needs to add the packet and let the game do the logic.

+++ src/ca/afroman/gui/GuiChooseRole.java
@@ -13,7 +13,7 @@ import ca.afroman.packet.PacketSetRole;
 
 public class GuiChooseRole extends GuiScreen
 {
-	private int playerID;
+	private short playerID;
 	private ConnectedPlayer player;
 	
 	private SpriteAnimation player1;
@@ -30,7 +30,7 @@ public class GuiChooseRole extends GuiScreen
 	private FlickeringLight light1;
 	private FlickeringLight light2;
 	
-	public GuiChooseRole(GuiScreen parentScreen, int playerID)
+	public GuiChooseRole(GuiScreen parentScreen, short playerID)
 	{
 		super(parentScreen);
