+++ src/ca/afroman/network/IPConnectedPlayer.java
@@ -16,7 +16,7 @@ public class IPConnectedPlayer extends ConnectedPlayer
 	 * @param role the role of the player
 	 * @param username the username of the player
 	 */
-	public IPConnectedPlayer(InetAddress address, int port, int id, Role role, String username)
+	public IPConnectedPlayer(InetAddress address, int port, short id, Role role, String username)
 	{
 		super(id, role, username);
