+++ src/ca/afroman/thread/DynamicTickRenderThread.java
@@ -5,7 +5,7 @@ public abstract class DynamicTickRenderThread extends DynamicTickThread
 	protected int frames;
 	protected int fps;
 	
-	public DynamicTickRenderThread(ThreadGroup group, String name, int ticksPerSecond)
+	public DynamicTickRenderThread(ThreadGroup group, String name, double ticksPerSecond)
 	{
 		super(group, name, ticksPerSecond);
 	}
@@ -22,7 +22,7 @@ public abstract class DynamicTickRenderThread extends DynamicTickThread
 	/**
 	 * Runs every time that the thread loops.
 	 */
-	// TODO is zynchronization starts to bread, make this synchronized
+	// TODO is zynchronization starts to bread, make this synchronised
 	@Override
 	public void onRun()
 	{
